[Essential]Extract FeedStoreSpy to its own file
